requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
celery==5.2.3
redis==4.0.2
scikit-learn==1.3.2
celery-redbeat==2.2.0